feat: 퀴즈 모달 기능 추가 및 Quiz 클래스 리팩토링
